+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRE2V02KNH2MQVYCPQ4PPAQP"
+        "01KREC7BX3T60HZ56DSF5TQZ4M"
     )
     r"""The unique identifier of the data source"""